feat: refactor announcement body component and add useAnnouncementToc hook for improved navigation
